==> 1-data_preprocessing.ipynb <==
# import packages
import pandas as pd
from sklearn.preprocessing import OneHotEncoder, StandardScaler

# add raw flights data to df
df = pd.read_csv('flights_raw.txt')
df.head()

# create day of week column from fl_date
print(type(df['fl_date'][0]))
df['fl_date'] = pd.to_datetime(df['fl_date'])
print(type(df['fl_date'][0]))
df['day_of_week'] = df['fl_date'].dt.dayofweek
set(df['day_of_week'])

# split into training and test
features = ['op_unique_carrier', 'crs_dep_time', 'crs_arr_time', 'crs_elapsed_time', 'distance', 'day_of_week']
X = df[features]
y = df['arr_delay']
X.shape, y.shape

### create preprocessed working file

# initialize scalers
scaler = StandardScaler()
enc = OneHotEncoder(sparse = False)

# OHE categorical features
cat = X['op_unique_carrier']
cat_array = cat.values.reshape(-1, 1)
cat_encoded = enc.fit_transform(cat_array)
cat_encoded

enc.categories_

enc.get_feature_names_out(['carrier'])

# scale numerical features
num = X.drop('op_unique_carrier', axis=1)
num_scaled = scaler.fit_transform(num)
num_scaled

# recombine processed files back into a single features file
cat_df = pd.DataFrame(cat_encoded, columns = enc.get_feature_names_out(['carrier']))
num_df = pd.DataFrame(num_scaled, columns = ['crs_dep_time', 'crs_arr_time', 'crs_elapsed_time', 'distance', 'day_of_week'])
X_scaled = pd.merge(num_df, cat_df, left_index = True, right_index = True)
X_scaled.head()

X_scaled.shape

# merge training and target file back to export
full_scaled_data = pd.merge(y, X_scaled, left_index = True, right_index = True)
full_scaled_data

# export
full_scaled_data.to_csv('full_scaled_data.csv')
