==> 2-model_selection.ipynb <==
# import packages
import pandas as pd
from sklearn import metrics as m
from sklearn.model_selection import train_test_split

import xgboost as xgb
from sklearn.svm import LinearSVR
from sklearn.ensemble import RandomForestRegressor, ExtraTreesRegressor, IsolationForest
from sklearn.ensemble import AdaBoostRegressor, GradientBoostingRegressor, HistGradientBoostingRegressor
from sklearn.ensemble import BaggingRegressor

# add data
df = pd.read_csv('full_scaled_data.csv', index_col = 0)
df.head(3)

# split training and target
X = df.drop('arr_delay', axis = 1)
y = df['arr_delay']

# replace NaN value with 0 (NaN means no arrival delay)
y = y.fillna(0)

# split data into training and test files
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

## model testing

# test LinearSVR with default values
svr = LinearSVR()

svr.fit(X_train, y_train)

y_pred_svr = svr.predict(X_test)

m.mean_squared_error(y_test, y_pred_svr)

# test AdaBoostRegressor with default values
ada = AdaBoostRegressor()

ada.fit(X_train, y_train)

y_pred_ada = ada.predict(X_test)

m.mean_squared_error(y_test, y_pred_ada)

# test GradientBoostingRegressor with default values
grad = GradientBoostingRegressor()

grad.fit(X_train, y_train)

y_pred_grad = grad.predict(X_test)

m.mean_squared_error(y_test, y_pred_grad)

# test HistGradientBoostingRegressor with default values
hist = HistGradientBoostingRegressor()

hist.fit(X_train, y_train)

y_pred_hist = hist.predict(X_test)

m.mean_squared_error(y_test, y_pred_hist)

# test BaggingRegressor with default values
bag = BaggingRegressor()

bag.fit(X_train, y_train)

y_pred_bag = bag.predict(X_test)

m.mean_squared_error(y_test, y_pred_bag)

# test IsolationForest with default values
iso = IsolationForest()

iso.fit(X_train, y_train)

y_pred_iso = iso.predict(X_test)

m.mean_squared_error(y_test, y_pred_iso)

# test XGBRegressor with default values
xgr = xgb.XGBRegressor()

xgr.fit(X_train, y_train)

y_pred_xgr = xgr.predict(X_test)

m.mean_squared_error(y_test, y_pred_xgr)

# test XGB RF Regressor with default values
xgf = xgb.XGBRFRegressor()

xgf.fit(X_train, y_train)

y_pred_xgf = xgf.predict(X_test)

m.mean_squared_error(y_test, y_pred_xgf)

# test RandomForestRegressor with default values
rnd = RandomForestRegressor(max_depth = 6)

rnd.fit(X_train, y_train)

y_pred_rnd = rnd.predict(X_test)

m.mean_squared_error(y_test, y_pred_rnd)

# test ExtraTreesRegressor with default values
xtr = ExtraTreesRegressor(max_depth = 6)

xtr.fit(X_train, y_train)

y_pred_xtr = xtr.predict(X_test)

m.mean_squared_error(y_test, y_pred_xtr)

# create results dictionary for MSE scores
results = {}
results['svr'] = 2402.5668964926836
results['adaboost'] = 92577.00331245457
results['gradientboost'] = 2270.6214976350693
results['hist g boost'] = 2259.206374858313
results['bagging'] = 2719.797988565089
results['isolation forest'] = 2309.0614013058766
results['xgb regressor'] = 2464.5472423371043
results['xgb rf regressor'] = 2291.2185461632534
results['random forest'] = 2287.0376456597637
results['extra trees'] = 2262.067272679809

# retry BaggingRegressor but PROPERLY
rebag = BaggingRegressor(LinearSVR())

rebag.fit(X_train, y_train)

y_pred_rebag = rebag.predict(X_test)

m.mean_squared_error(y_test, y_pred_rebag)

results['bagged svr'] = 2402.9677356422662

results

# manual sorting by scores, ascending order of accuracy and picking models
{
 'adaboost': 92577.00331245457,
 # 'bagging': 2719.797988565089,
 # 'xgb regressor': 2464.5472423371043,
 # 'bagged svr': 2402.9677356422662,
 # 'svr': 2402.5668964926836,
 'isolation forest': 2309.0614013058766,
 # 'xgb rf regressor': 2291.2185461632534,
 # 'random forest': 2287.0376456597637,
 # 'gradientboost': 2270.6214976350693,
 'extra trees': 2262.067272679809,
 'hist g boost': 2259.206374858313
}

### r-squared scores

# LinearSVR with default values
m.r2_score(y_test, y_pred_svr)

# AdaBoostRegressor with default values
m.r2_score(y_test, y_pred_ada)

# GradientBoostingRegressor with default values
m.r2_score(y_test, y_pred_grad)

# HistGradientBoostingRegressor with default values
m.r2_score(y_test, y_pred_hist)

# BaggingRegressor with default values
m.r2_score(y_test, y_pred_bag)

# IsolationForest with default values
m.r2_score(y_test, y_pred_iso)

# XGBRegressor with default values
m.r2_score(y_test, y_pred_xgr)

# XGB RF Regressor with default values
m.r2_score(y_test, y_pred_xgf)

# RandomForestRegressor with default values
m.r2_score(y_test, y_pred_rnd)

# ExtraTreesRegressor with default values
m.r2_score(y_test, y_pred_xtr)

# BaggingRegressor using LinearSVR with default values
m.r2_score(y_test, y_pred_rebag)

### adjusted r-squared scores

X.shape, y.shape

# LinearSVR with default values
1 - (1-m.r2_score(y_test, y_pred_svr)) * (len(y_train)-1) / (len(y_train)-33-1)

# AdaBoostRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_ada)) * (len(y_train)-1) / (len(y_train)-33-1)

# GradientBoostingRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_grad)) * (len(y_train)-1) / (len(y_train)-33-1)

# HistGradientBoostingRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_hist)) * (len(y_train)-1) / (len(y_train)-33-1)

# BaggingRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_bag)) * (len(y_train)-1) / (len(y_train)-33-1)

# IsolationForest with default values
1 - (1-m.r2_score(y_test, y_pred_iso)) * (len(y_train)-1) / (len(y_train)-33-1)

# XGBRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_xgr)) * (len(y_train)-1) / (len(y_train)-33-1)

# XGB RF Regressor with default values
1 - (1-m.r2_score(y_test, y_pred_xgf)) * (len(y_train)-1) / (len(y_train)-33-1)

# RandomForestRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_rnd)) * (len(y_train)-1) / (len(y_train)-33-1)

# ExtraTreesRegressor with default values
1 - (1-m.r2_score(y_test, y_pred_xtr)) * (len(y_train)-1) / (len(y_train)-33-1)

# BaggingRegressor using LinearSVR with default values
1 - (1-m.r2_score(y_test, y_pred_rebag)) * (len(y_train)-1) / (len(y_train)-33-1)
