==> 3-parameter_tuning.ipynb <==
# import packages
import pandas as pd
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import ExtraTreesRegressor, HistGradientBoostingRegressor

# add data
df = pd.read_csv('full_scaled_data.csv', index_col = 0)
df.head(3)

# split data into target and training
X = df.drop('arr_delay', axis = 1)
y = df['arr_delay']
X.shape, y.shape

# replace nan values in target with 0 (i.e. no delay)
y = y.fillna(0)
y.isnull().sum()

### parameter tuning for ExtraTreesRegressor using GridSearchCV

xtr_params = {
    'n_estimators': [50, 100, 200],
    'max_depth': [3, 6, 9]    
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [100, 200, 300],
    'max_depth': [3, 6, 7]    
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [150, 200, 250],
    'max_depth': [3, 6, 7]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [180, 200, 220],
    'max_depth': [4, 6]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [190, 200, 210],
    'max_depth': [5, 6]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [195, 200, 205],
    'max_depth': [5, 6]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [198, 200, 202]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(max_depth = 5), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [196, 197, 198, 199]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(max_depth = 5), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'max_depth': [5, 6]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(n_estimators = 197), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [196, 197, 198],
    'max_depth': [4, 5, 6]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_params = {
    'n_estimators': [196, 197, 198],
    'max_depth': [5, 6, 7]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_grid.best_score_

xtr_params = {
    'max_depth': [5, 6],
    'warm_start': [True, False]
}
xtr_grid = GridSearchCV(ExtraTreesRegressor(n_estimators = 197), xtr_params)

xtr_grid.fit(X, y)

xtr_grid.best_estimator_

xtr_grid.best_params_

xtr_grid.best_score_

### parameter tuning for HistGradientBoostingRegressor using GridSearchCV

hist_params = {
    'learning_rate': [0.01, 0.1, 0.3],
    'max_depth': [3, 6, 9],
    'l2_regularization': [0, 0.1]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_estimator_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.05, 0.1, 0.2],
    'max_depth': [3, 4, 5]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.01, 0.05, 0.1],
    'max_depth': [2, 3, 4]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.01, 0.03, 0.05, 0.07, 0.09, 0.11, 0.13, 0.15, 0.2],
    'max_depth': [2, 3, 4]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.05, 0.06, 0.07, 0.08, 0.09, 0.1],
    'max_depth': [2, 3, 4],
    'l2_regularization': [0, 0.1],
    'warm_start': [True, False]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.07, 0.08, 0.09, 0.1, 0.13, 0.16],
    'max_depth': [2, 3, 4],
    'l2_regularization': [0, 0.01, 0.1, 0.2],
    'warm_start': [True, False]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.07, 0.1, 0.13, 0.16, 0.19],
    'max_depth': [2, 3, 4],
    'l2_regularization': [0, 0.05, 0.1, 0.15],
    'warm_start': [True, False]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.07, 0.1, 0.13, 0.16, 0.19],
    'max_depth': [2, 3, 4],
    'l2_regularization': [0, 0.05, 0.1, 0.15],
    'warm_start': [True, False]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(), hist_params)

hist_grid.fit(X, y)

hist_grid.best_params_

hist_grid.best_score_

hist_grid.fit(X, y)

hist_grid.best_params_

hist_grid.best_score_

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.07, 0.1, 0.13, 0.16, 0.19],
    'l2_regularization': [0, 0.05, 0.1, 0.15]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.07, 0.10, 0.13, 0.16, 0.19],
    'l2_regularization': [0.05, 0.10, 0.15]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.17, 0.19, 0.21],
    'l2_regularization': [0.10, 0.15, 0.18]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.17, 0.19, 0.21],
    'l2_regularization': [0.10, 0.12, 0.15]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.17, 0.19, 0.21],
    'l2_regularization': [0.12, 0.15]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.17, 0.19, 0.21],
    'l2_regularization': [0.12, 0.15],
    'max_iter': [100, 200]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_params = {
    'learning_rate': [0.17, 0.19, 0.21, 0.23, 0.25],
    'l2_regularization': [0.12, 0.13, 0.14, 0.15]
}
hist_grid = GridSearchCV(HistGradientBoostingRegressor(warm_start = True, max_depth = 3), hist_params)

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_

hist_grid.fit(X, y)

hist_grid.best_score_

hist_grid.best_params_
