==> 4-final.ipynb <==
# import libraries
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import HistGradientBoostingRegressor, ExtraTreesRegressor, StackingRegressor

## file processing and feature engineering

pd.options.display.max_columns = None
pd.options.display.max_rows = None

#load dataset
flights = pd.read_csv('flights_raw.txt')

#feature engineering
flights['fl_date'] = pd.to_datetime(flights['fl_date'], format='%Y-%m-%d')
flights['year'] = flights['fl_date'].dt.year
flights['month'] = flights['fl_date'].dt.month
flights['day'] = flights['fl_date'].dt.day
flights['arr_delay'] = flights['arr_delay'].fillna(0)
flights = flights.dropna()

#Dropping the outlier rows with standard deviation
factor = 3
upper_lim = flights['arr_delay'].mean() + flights['arr_delay'].std() * factor
lower_lim = flights['arr_delay'].mean() - flights['arr_delay'].std() * factor

flights = flights[(flights['arr_delay'] < upper_lim) & (flights['arr_delay'] > lower_lim)]

#binning distance
flights['bin_dist'] = pd.cut(flights['distance'], bins=[0,500,1000,1500,2000,2500,3000,3500,4000,4500,5000], labels=[1,2,3,4,5,6,7,8,9,10])

#binning time
flights['bin_time'] = pd.cut(flights['crs_elapsed_time'], bins=[0,120,240,360,480,600,2000], labels=[1,2,3,4,5,6])

#log of distance
flights['log_dist'] = (flights['distance']+1).transform(np.log)

#log of elapsed time
flights['log_time'] = (flights['crs_elapsed_time']+1).transform(np.log)

#one-hot encoding for op_unique_carrier
encoded_columns = pd.get_dummies(flights['op_unique_carrier'])
flights = flights.join(encoded_columns).drop('op_unique_carrier', axis=1)

flights.head()

# add test file
test_df = pd.read_csv('test-jan-1_7.csv')
test_df.head(3)

## processing the test file

# feature engineer test file

#feature engineering
test_df['fl_date'] = pd.to_datetime(test_df['fl_date'], format='%Y-%m-%d')
test_df['year'] = test_df['fl_date'].dt.year
test_df['month'] = test_df['fl_date'].dt.month
test_df['day'] = test_df['fl_date'].dt.day

#binning distance
test_df['bin_dist'] = pd.cut(test_df['distance'], bins=[0,500,1000,1500,2000,2500,3000,3500,4000,4500,5000], labels=[1,2,3,4,5,6,7,8,9,10])

#binning time
test_df['bin_time'] = pd.cut(test_df['crs_elapsed_time'], bins=[0,120,240,360,480,600,2000], labels=[1,2,3,4,5,6])

#log of distance
test_df['log_dist'] = (test_df['distance']+1).transform(np.log)

#log of elapsed time
test_df['log_time'] = (test_df['crs_elapsed_time']+1).transform(np.log)

#one-hot encoding for op_unique_carrier
encoded_columns = pd.get_dummies(test_df['op_unique_carrier'])
test_df = test_df.join(encoded_columns).drop('op_unique_carrier', axis=1)

test_df.head()

# strip relevant columns from test file
feat = ['op_unique_carrier', 'crs_dep_time', 'crs_arr_time', 'crs_elapsed_time', 'distance', 'day_of_week']
test = test_df[feat]
test.shape

## validation

# Split-out validation dataset
features = ['origin_airport_id','dest_airport_id','crs_dep_time','log_time','month','log_dist','9E','AA','AS','AX','B6','C5','CP','DL','EM','EV','F9','G4','G7','HA','MQ','NK','OH','OO','PT','QX','UA','WN','YV','YX','ZW']
X = flights[features]
y = flights['arr_delay']
X_train, X_test, Y_train, Y_test = train_test_split(X, y, test_size=0.20, random_state=1)

test = test_df[features]

flights[features].to_csv('preprocessed_data.csv', index=False)

Y_train = Y_train.values.reshape(-1, 1)
Y_test = Y_test.values.reshape(-1, 1)
X_train.shape, X_test.shape, Y_train.shape, Y_test.shape

## final model training and prediction

# add models with determined parameters to list for meta-estimator
estimators = [
    ('trees', ExtraTreesRegressor()),
    ('boost', HistGradientBoostingRegressor(
        warm_start = True
    ))
]

# initialize meta estimator
reg = StackingRegressor(
    estimators = estimators,
    passthrough = True)

# train model on full sampled data
reg.fit(X_train, Y_train)

# training score
reg.score(X_train, Y_train)

# predict on submission test data
y_pred = reg.predict(test)
y_pred

## create final submission file

# add sample submission file to copy structure
sample = pd.read_csv('mid-term-project-I/sample_submission.csv')
sample.head(3)

# get column names from sample file
cols = sample.columns.tolist()
cols

# remove prediction column name
cols.remove('predicted_delay')

# create submission df from features on submission test file using column names from sample
submission = test_df[cols]
submission.head()

# add predictions to submission file
submission['predicted_delay'] = y_pred
submission.head()

# export final submission file
submission.to_csv('submission.csv', index=False)
